use lodash with calendar
