+++ package.json
@@ -19,6 +19,7 @@
     "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.14.0",
     "eslint-plugin-react-hooks": "^2.1.1",
+    "lodash": "^4.17.15",
     "lru-cache": "^5.1.1",
     "mdbreact": "^4.21.0",
     "moment": "latest",
